+++ src/components/Modal/Body.jsx
@@ -3,10 +3,10 @@ import React from "react";
 import classnames from "classnames";
 import PropTypes from "prop-types";
 
-const Body = ({ children, className, dataTestId }) => (
+const Body = ({ children, className, dataTestid }) => (
   <div
     className={classnames("neeto-ui-modal__body", className)}
-    data-testid={dataTestId ?? "modal-body"}
+    data-testid={dataTestid ?? "modal-body"}
   >
     {children}
   </div>