+++ src/components/Modal/Footer.jsx
@@ -3,10 +3,10 @@ import React from "react";
 import classnames from "classnames";
 import PropTypes from "prop-types";
 
-const Footer = ({ children, className, dataTestId }) => (
+const Footer = ({ children, className, dataTestid }) => (
   <div
     className={classnames("neeto-ui-modal__footer", className)}
-    data-testid={dataTestId ?? "modal-footer"}
+    data-testid={dataTestid ?? "modal-footer"}
   >
     {children}
   </div>