+++ src/layouts/Sidebar/ProfileSection/index.jsx
@@ -4,13 +4,13 @@ import React, { forwardRef } from "react";
 import Avatar from "components/Avatar";
 
 const ProfileSection = forwardRef(({ profileInfo, onClick }, ref) => {
-  const dataTestId = profileInfo["data-testid"] || "profile-section";
+  const dataTestid = profileInfo["data-testid"] || "profile-section";
 
   return (
     <button
       {...{ onClick, ref }}
       className="neeto-ui-w-full neeto-ui-text-left neeto-ui-sidebar__profile-wrapper"
-      data-testid={dataTestId}
+      data-testid={dataTestid}
     >
       <span className="neeto-ui-flex neeto-ui-items-center neeto-ui-flex-shrink-0 neeto-ui-w-full neeto-ui-sidebar__profile">
         <Avatar